utils/requirements.txt
openai>=1.3.2
nltk
datasets
sympy
python-chess
